requirements.txt
flask
flask-cors
dataset
psycopg2-binary
gunicorn
python-slugify
marshmallow
